Move AuthorizationBearer to new namespace
